fix datatype syntax highlighting (#153)
Agda returns 'datatype', not 'type', as the name of datatypes.
CornelisType in Pretty.hs will not apply to these (via toAtomName)
unless it's named CornelisDatatype.

This is the simplest fix. One could add a CornelisType syntax group that
both CornelisDatatype and CornelisRecord link to, which might be a
little more backwards-compatible with people who do highlight linking
around CornelisType. But my guess is that CornelisType has never worked
(see #153) and so I'm not worrying about backwards compatibility.